Check permission b4 create/delete category

+++ src/main/java/com/lifecheatsheet/springbackend/services/CategoryService.java
@@ -3,6 +3,7 @@ package com.lifecheatsheet.springbackend.services;
 import com.lifecheatsheet.springbackend.dtos.CategoryCreateDto;
 import com.lifecheatsheet.springbackend.entities.Category;
 import com.lifecheatsheet.springbackend.entities.User;
+import com.lifecheatsheet.springbackend.exception.ForbiddenException;
 import com.lifecheatsheet.springbackend.exception.NotFoundException;
 import com.lifecheatsheet.springbackend.repositories.CategoryRepository;
 import com.lifecheatsheet.springbackend.repositories.UserRepository;
@@ -21,20 +22,25 @@ public class CategoryService {
     @Autowired
     UserRepository userRepository;
 
-    public List<Category> getAllCategories(){
+    @Autowired
+    PermissionService permissionService;
+
+    public List<Category> getAllCategories() {
         return categoryRepository.findAll();
     }
 
     public Category getCategoryById(int id) throws NotFoundException {
         Optional<Category> categoryOptional = categoryRepository.findById(id);
-        if(categoryOptional.isEmpty())
+        if (categoryOptional.isEmpty())
             throw new NotFoundException("This category cannot be found");
 
         return categoryOptional.get();
     }
 
-    public Category createNewCategory(CategoryCreateDto newCategory, String userEmail){
+    public Category createNewCategory(CategoryCreateDto newCategory, String userEmail) throws ForbiddenException {
         User author = userRepository.findByEmail(userEmail);
+
+        permissionService.checkCreateAndDeleteCategoryPermission(author);
 
         Category category = new Category();
         category.setCreatedAt(new Date());
@@ -44,7 +50,11 @@ public class CategoryService {
         return categoryRepository.saveAndFlush(category);
     }
 
-    public void deleteCategory(int id) throws NotFoundException {
+    public void deleteCategory(int id, String userEmail) throws NotFoundException, ForbiddenException {
+        User author = userRepository.findByEmail(userEmail);
+
+        permissionService.checkCreateAndDeleteCategoryPermission(author);
+
         Category category = getCategoryById(id);
         categoryRepository.delete(category);
     }
